refactor: re-design member form
